% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{add_ratios}
\alias{add_ratios}
\title{Add ratios between spatio-temporal neighborhood indicators}
\usage{
add_ratios(df, var, indStat = "mean")
}
\arguments{
\item{df}{A data frame of spatio-temporal indicators. 
Column names should be of type <variable name>_<indStat>_<radius>.}

\item{var}{A character string with the name of the variable
with indicators to add ratios}

\item{indStat}{The name of the summarizing stat that was used
to calculate the indicators}
}
\value{
A data frame including the original data of \code{df} 
and additional ratios between the indicators of subsequent radiuses.
}
\description{
Add ratios between spatio-temporal neighborhood indicators
}
\seealso{
\code{\link{get_st_indicators}}
}
