% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics.R
\name{ae}
\alias{ae}
\alias{se}
\alias{mse}
\alias{rmse}
\alias{mae}
\alias{nmae}
\alias{nmse}
\alias{nrmse}
\title{Calculate absolute error}
\usage{
ae(y, y_hat)

se(y, y_hat)

mse(y, y_hat, na.rm = TRUE)

rmse(y, y_hat, na.rm = TRUE)

mae(y, y_hat, na.rm = TRUE)

nmae(y, y_hat, y_train = NULL, statFUN = median, na.rm = TRUE)

nmse(y, y_hat, y_train = NULL, statFUN = mean, na.rm = TRUE)

nrmse(y, y_hat, y_train = NULL, statFUN = mean, na.rm = TRUE)
}
\arguments{
\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{y_train}{a vector of training values}

\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{y_train}{a vector of training values}

\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{y_train}{a vector of training values}
}
\value{
a vector of absolute errors

a vector of squared errors

the mean squared error

a vector of root mean squared errors

the mean absolute error

the mean absolute error

the mean absolute error

the mean absolute error
}
\description{
Calculate absolute error

Calculate squared error

Calculate mean squared error

Calculate root mean squared error

Calculate mean absolute error

Calculate normalized mean absolute error

Calculate normalized mean squared error

Calculate normalized root mean squared error
}
\section{Functions}{
\itemize{
\item \code{ae}: 

\item \code{se}: 

\item \code{mse}: 

\item \code{rmse}: 

\item \code{mae}: 

\item \code{nmae}: 

\item \code{nmse}: 

\item \code{nrmse}: 
}}

