% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics.R
\name{ae}
\alias{ae}
\alias{se}
\title{Calculate vector of error values}
\usage{
ae(y, y_hat)

se(y, y_hat)
}
\arguments{
\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}
}
\value{
a vector of error values
}
\description{
Calculate vector of error values
}
\section{Functions}{
\itemize{
\item \code{ae}: absolute error

\item \code{se}: squared error
}}

