% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{compressAllRes}
\alias{compressAllRes}
\title{Compress results from all (artificial) experiments}
\usage{
compressAllRes(all.res, rmAllRaw = F)
}
\arguments{
\item{all.res}{a list with multiple levels (model,
grid size, series size and lists of full results of multiple experiments)}

\item{rmAllRaw}{a boolean indicating whether the whole rawRes should
be removed (defaults to FALSE). If TRUE, only "train" data will be
removed from each set of results}
}
\value{
A multi-level list containing compressed results. Either all rawRes
is removed, or \code{train} is substituted by a vector of the number of instances, 
time and location IDs in the training set, in both \code{out_estRes} and
\code{in_estRes}.
}
\description{
Compress results from all (artificial) experiments
}
\seealso{
\code{\link{summarize_all_exps}}, \code{\link{run_all_experiments}}
}
