% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{compressRes}
\alias{compressRes}
\title{Compress results from one experiment}
\usage{
compressRes(res, rmAllRaw = F)
}
\arguments{
\item{res}{A list containing full results of one experiment
(out_estRes and in_estRes)}

\item{rmAllRaw}{a boolean indicating whether the whole rawRes should
be removed (defaults to FALSE). If TRUE, only \code{train} data will be
substituted by a vector of the number of instances, time and location IDs 
in the training set}
}
\value{
A list containing compressed results of one experiment. 
Either all rawRes is removed, or \code{train} substituted by a vector 
of the number of instances, time and location IDs in the training set
in both \code{out_estRes} and \code{in_estRes}
}
\description{
Compress results from one experiment
}
\seealso{
\code{\link{summarize_one_exp}}, \code{\link{run_one_experiment}}
}
