% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_folds}
\alias{cv_folds}
\title{Cut into folds}
\usage{
cv_folds(x, nfolds)
}
\arguments{
\item{x}{a data.frame}

\item{nfolds}{number of folds}
}
\value{
a vector with the fold assignment of each row
}
\description{
Assigns rows of a data frame into folds for cross-validation.
}
