% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{embed_series}
\alias{embed_series}
\title{Embed each time series in a spatio-temporal data set}
\usage{
embed_series(df, var, k, time = "time", station_id = "station")
}
\arguments{
\item{df}{data frame}

\item{var}{a character string, the name of the variable to embed}

\item{k}{a numeric, the embed size}

\item{time}{a character string, the column name identifying
the time of observation}

\item{station_id}{a character string, the column name
identifying the location of observation}
}
\value{
A data frame with extra columns var_Tm1, var_Tm2, ...,
var_Tm\(k-1\)
}
\description{
Embed each time series in a spatio-temporal data set
}
