% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starma_utils.R
\name{exp_c}
\alias{exp_c}
\title{Exponential function}
\usage{
exp_c(x, C = 10000)
}
\arguments{
\item{x}{A vector}

\item{C}{A constant}
}
\value{
A vector \code{exp(-x/C)}
}
\description{
Calculate \code{exp(-x/C)}.A list of matrices like the ones returned by consecutive use of 
\code{spdep::dnearneigh} and \code{spdep::nblag} where a value higher than 0
 implies that the row and column locations are neighbours
}
