% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starma_utils.R
\name{generate_coef}
\alias{generate_coef}
\title{Generate coefficients for a STARMA stationary process}
\usage{
generate_coef(coef_specs = list(c_10 = c(-2, 2), c_11 = c(-2, 2), c_20 =
  c(-1, 1), c_21 = 0), type = "STAR", ndigits = 3)
}
\arguments{
\item{coef_specs}{A named list of coefficient specifications for the models.
Each slot in the list should itself contain a named list
with slots \code{c_10}, \code{c_11}, \code{c_20}, \code{c_21} 
containing either a number which will be used directly for the 
coefficient of that order in the AR and/or the MA components of 
the model, or a vector specifying an interval whithin which a 
coefficient will be randomly generated.}

\item{type}{A vector of the types of STARMA models to be used for data
generation. Should be \code{STAR},\code{STARMA},\code{NL_STAR} or \code{STMA}.}

\item{ndigits}{Number of digits to keep when rounding coefficients}
}
\value{
A vector of stationary coefficients generated with names 
\code{phi_10}, \code{phi_11}, \code{phi_20}, \code{phi_21},
\code{theta_10}, \code{theta_11}, \code{theta_20}, \code{theta_21} and \code{FUN}.
}
\description{
Generate coefficients for a STARMA stationary process
}
\seealso{
\code{\link{starma_stat_check}}
}
