% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_utils.R
\name{generate_grid}
\alias{generate_grid}
\title{Create a regular grid using spdep package}
\usage{
generate_grid(Nsites, grid.h = ceiling(sqrt(Nsites)),
  grid.w = ceiling(sqrt(Nsites)))
}
\arguments{
\item{Nsites}{Number of locations in the grid}

\item{grid.h}{Height of the grid (in number of locations).
Defaults to \code{sqrt(Nsites)}}

\item{grid.w}{Width of the grid (in number of locations).
Defaults to \code{sqrt(Nsites)}}
}
\value{
A list with slots sites -- a data frame with
columns id, x and y --, and klist -- a list of neighbour
matrices of order 0 and 1 where the values are calculated
using functions from \code{spdep} package.
}
\description{
Create a regular grid with a certain number of locations,
returning each location's position and neighbour matrices
of order 0 and 1.
}
\seealso{
\code{\link{dnearneigh}}, \code{\link{nblag}},
\code{\link{spdep::nb2mat}}
}
