% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_utils.R
\name{generate_multiple_datasets}
\alias{generate_multiple_datasets}
\title{Generate multiple artificial datasets using STARMA on regular grids}
\usage{
generate_multiple_datasets(Nsites, Ntimes, mtypes, coef_specs, ncoefs,
  trash = 0, grid.h = ceiling(sqrt(Nsites)),
  grid.w = ceiling(sqrt(Nsites)), init_seed = 1234, mid_seed = NULL,
  sim_seed = NULL)
}
\arguments{
\item{Nsites}{A vector containing the number of locations in the regular
grids to be generated}

\item{Ntimes}{A vector containing the number of time series points to be
generated for each location}

\item{mtypes}{A vector of the types of STARMA models to be used for data
generation. Should be \code{STAR},\code{STARMA},\code{NL_STAR} or \code{STMA}.}

\item{coef_specs}{A named list of coefficient specifications for the models.
Each slot in the list should itself contain a named list
with slots \code{c_10}, \code{c_11}, \code{c_20}, \code{c_21} 
containing either a number which will be used directly for the 
coefficient of that order in the AR and/or the MA components of 
the model, or a vector specifying an interval whithin which a 
coefficient will be randomly generated. Coefficients are only randomly generated
for the first pair of \code{Nsites} and \code{Ntimes}, and then are re-used
for different grid and time series sizes.}

\item{ncoefs}{A vector with the number of data sets that should be generated
according to each of the specifications in the \code{coef_specs} list.}

\item{trash}{A number of initial time series points to be discarded at 
each location.}

\item{grid.h}{Height of the grid (in number of locations).
Defaults to \code{sqrt(Nsites)}.}

\item{grid.w}{Width of the grid (in number of locations).
Defaults to \code{sqrt(Nsites)}.}

\item{init_seed}{Seed to set at the begining (before coefficient generation).
Default is \code{1234}.}

\item{sim_seed}{Seed to feed to \code{generate_one_dataset}.
Default is \code{NULL}.}

\item{init_seed}{Seed to set between grid/time series size change.
Default is \code{NULL}.}
}
\value{
A list containing a list for each size in \code{Nsites}. Each
of the sublists contains a list of datasets generated by 
\code{generate_one_dataset} for each time series size specified in 
\code{Ntimes}. Each of these sublists will have 
\code{sum(length(mtypes) x ncoefs)} datasets of  each grid and
 time series size. In total, \code{length(Nsites) x length(Ntimes) 
 x sum(length(mtypes) x ncoefs)} spatio-temporal datasets are generated.
}
\description{
This function generates multiple regular grids and simulates time series
for each location according to one or more STARMA model specifications.
}
\seealso{
\code{\link{generate_one_dataset}}
}
