% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_utils.R
\name{generate_one_dataset}
\alias{generate_one_dataset}
\title{Generate an artificial dataset using STARMA on a regular grid}
\usage{
generate_one_dataset(Nsites, Ntimes, coef_spec, gen_coef = TRUE,
  mtype = NULL, trash = 0, ndigits = 3, seed = NULL,
  grid.h = ceiling(sqrt(Nsites)), grid.w = ceiling(sqrt(Nsites)))
}
\arguments{
\item{Nsites}{Number of locations in the grid}

\item{Ntimes}{Number of points in each time series to generate}

\item{coef_spec}{Coefficient specifications. Should be a named list
with slots \code{c_10}, \code{c_11}, \code{c_20}, \code{c_21} 
containing either a number which will be used directly for the 
coefficient of that order in the AR and/or the MA components of 
the model, or a vector specifying an interval whithin which a 
coefficient will be randomly generated.}

\item{gen_coef}{A Boolean indicating whether \code{coef_spec} should be used
to generate coefficients (TRUE: default) or whether \code{coef_spec} already
contains a proper list of coefficients}

\item{mtype}{Type of STARMA model (should only be provided if \code{gen_coef} 
is \code{TRUE}. One of \code{STAR},\code{STARMA},\code{NL_STAR} or \code{STMA}.}

\item{trash}{A number of initial time series points to be discarded
at each location.}

\item{ndigits}{Number of digits to keep when rounding coefficients.}

\item{seed}{Seed to feed to \code{generate_stdata}. Default is \code{NULL}.}

\item{grid.h}{Height of the grid (in number of locations).
Defaults to \code{sqrt(Nsites)}}

\item{grid.w}{Width of the grid (in number of locations).
Defaults to \code{sqrt(Nsites)}}
}
\value{
A list with three slots: \code{data}, a matrix of spatio-temporal data 
(columns correspond to locations, rows to time-stamps); \code{coef}, a vector of 
the coefficients actually used to generate the data; and \code{grid}, a list
containing the positions of location and the neighbour matrices.
 
 @seealso \code{\link{generate_stdata}}, \code{\link{generate_coef}}, \code{\link{generate_grid}}
}
\description{
This function generates a regular grid and simulates a time series
for each location according to a STARMA model specification.
}
