% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starma_utils.R
\name{generate_stdata}
\alias{generate_stdata}
\title{Simulate a STARMA stationary process}
\usage{
generate_stdata(Ntimes, klist, coef, scale = FALSE, trash = 100,
  seed = NULL)
}
\arguments{
\item{Ntimes}{Length of time series to generate}

\item{klist}{A list of matrices like the ones returned by consecutive use of 
\code{spdep::dnearneigh} and \code{spdep::nblag} where a value higher than 0
 implies that the row and column locations are neighbours}

\item{coef}{A named vector of stationary stationary coefficients with names 
\code{phi_10}, \code{phi_11}, \code{phi_20}, \code{phi_21},
\code{theta_10}, \code{theta_11}, \code{theta_20}, \code{theta_21} and \code{FUN}.}

\item{scale}{A boolean indicating whether the data should be scaled around 0.}

\item{trash}{Number of initial time series points to discard}

\item{seed}{Seed to feed to \code{starma_sim}. Default is \code{NULL}.}
}
\value{
A list with two slots: \code{coef}, a vector of the coefficients
 actually used to generate the data; \code{data}, a matrix
 of spatio-temporal data (columns correspond to locations, rows to
 time-stamps).
 
 @seealso \code{\link{starma_sim}}
}
\description{
Simulate a STARMA stationary process
}
