% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{get_all_neib_vals}
\alias{get_all_neib_vals}
\title{Get the spatio-temporal neighbourhoods of
all observations}
\usage{
get_all_neib_vals(df, max_radius, t_dist_mat, s_dist_mat, alpha, vars, time_id,
  site_id, parallel = FALSE, nsplits = 4)
}
\arguments{
\item{df}{a data frame of observations}

\item{max_radius}{the maximum spatio-temporal distance
allowed to be included in a neighbourhood}

\item{t_dist_mat}{a matrix of normalized temporal distances
between time-stamps (rownames and colnames should be
a concatenation of "TIME_" and the time-stamp)}

\item{s_dist_mat}{a matrix of normalized spatial distances
between locations (rownames and colnames should be
a concatenation of "SITE_" and the location IDs)}

\item{alpha}{a weighting factor for the spatio-temporal distance}

\item{time_id}{the name of the column containing time-stamps}

\item{site_id}{the name of the column containing location IDs}

\item{parallel}{Boolean indicating whether the code should
run in parallel. Default is FALSE}

\item{nsplits}{Number of subsets of rows to split the data 
frame into so they can be processed in parallel}
}
\value{
A data frame where each row describes a neighbour,
with the first two columns containing the location ID and
time-stamp of the central observation, followed by two
columns with the neighbouring location ID and time-stamp,
a colunm containing the spatio-temporal distance between 
the two, and a final column containing the values of 
the variables in df at the neighbouring time and location.
}
\description{
Get the spatio-temporal neighbourhoods of
all observations
}
\details{
The spatio-temporal distance is defined as
\deqn{D_{i,j} = d_{i,j} x \alpha + t_{i,j}
x (1-\alpha)}
where d_{i,j} is the spatial distance between
locations, {t_{i,j}} is the temporal distance
between time-stapms and \eqn{\alpha} is a weighting factor.
Note that the \code{radius} should always be a number between
zero and \code{min(alpha, alpha-1)}.
Also note that if \code{alpha} is set to 1, then instead of a cone, 
the neighbourhood will have the shape of a cylinder.
}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.258.8742&rep=rep1&type=pdf}
}
\seealso{
\code{\link{get_st_neighbours}}
}
