% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{get_full_indicators}
\alias{get_full_indicators}
\title{Get time series embeds and spatio-temporal indicators}
\usage{
get_full_indicators(df, stations, k, betas, alpha = 0.5, var = "value",
  stats = c("mean", "weighted.mean", "sd"), ratios2add = c(TRUE, TRUE,
  FALSE), neib_type = "cone", parallel = FALSE, nsplits = 1,
  time_id = "time", site_id = "station")
}
\arguments{
\item{df}{A data frame containing spatio-temporal
information}

\item{k}{A numeric indicating the temporal embed size \(number\)}

\item{alpha}{a weighting factor for the spatio-temporal distance}

\item{var}{The name of the variable to summarize into indicators}

\item{stats}{A vector containing the names of functions
that are to be used to calculate summarizing statistics}

\item{ratios2add}{A vector of Boolean values indicating, for each
statistic in \code{stats} whether ratios between neighorhoods of 
subsequent sizes should be included as extra columns}

\item{neib_type}{the type of neighborhood to consider.
Can be
\itemize{
\item \code{cone} (default) - a cone with the center of its base at 
the observation (spatial radius growing with time)
\item \code{reversed} - a cone with its peak at the observation
(spatial radius shrinking with time)
}}

\item{parallel}{Boolean indicating whether the code should
run in parallel. Default is FALSE}

\item{nsplits}{Number of subsets of rows to split the data 
frame into so they can be processed in parallel}

\item{time_id}{The name of the column containing 
time-stamps in \code{df}}

\item{site_id}{The name of the column containing 
location IDs in \code{df}}
}
\value{
A data frame that contains extra columns <var>_Tm1, <var>_Tm2, 
..., <var>_Tm<k-1> with previous observations for each location,
summary statistics of the values of \code{var} found within the 
spatio-temporal neighbourhoods of the one or more \code{radiuses} 
of each pair (location ID, time-stamp) and ratios between them
}
\description{
Get time series embeds and spatio-temporal indicators
}
