% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{get_spatial_dist_mat}
\alias{get_spatial_dist_mat}
\title{Calculate spatial distance matrix}
\usage{
get_spatial_dist_mat(sites_sf, site_id)
}
\arguments{
\item{sites_sf}{an \code{sf} object with the geograhic
information of the locations (as returned by \code{df2site_sf})}

\item{site_id}{the column name of the location ID}
}
\value{
a matrix of distances. Row and column names
are a concatenation of "SITE_" and the location IDs.
}
\description{
A function that calculates the geographical distance
matrix between the locations of an \code{sf} object.
}
\seealso{
\code{\link{df2site_sf}}
}
