% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{get_st_indicator}
\alias{get_st_indicator}
\title{Get a spatio-temporal indicator from neighbourhood
data frame}
\usage{
get_st_indicator(all_neib_vals, stat, radius, ind_name, var,
  time_id = "time", site_id = "site")
}
\arguments{
\item{all_neib_vals}{a data frame containing
information on observations spatio-temporal distance
to neighbours and variable values at the neighbouring
locations and times.}

\item{stat}{the name of a function that calculates a
statistic (e.g., "mean"). If the stat is "weighted.mean"
then the inverse of the spatio-temporal distance is used 
to weight the values of observations in the spatio-temporal 
neighbourhood}

\item{radius}{a value defining the maximum spatio-temporal
distance allowed for an observation to be considered
within a spatio-temporal neighbourhood}

\item{ind_name}{the name of the indicator column}

\item{var}{the name of the variable to summarize into 
an indicator}

\item{time_id}{the name of the column containing 
time-stamps}

\item{site_id}{the name of the column containing 
location IDs}
}
\value{
A data frame that contains a summary statistic
of the values found within the spatio-temporal neighbourhood
of a certain radius of each pair (location ID, time-stamp) in
\code{all_neib_vals}
}
\description{
Calculate a spatio-temporal indicator of a certain
variable within a spatio-temporal neighborhood of a 
certain radius.
}
\details{
The spatio-temporal distance is defined as
\deqn{D_{i,j} = d_{i,j} x \alpha + t_{i,j}
x (1-\alpha)}
where d_{i,j} is the spatial distance between
locations, {t_{i,j}} is the temporal distance
between time-stapms and \eqn{\alpha} is a weighting factor.
Note that \code{radius} should always be a number between
zero and \code{min(alpha, alpha-1)}, so the border conditions apply.
Also note that if \code{alpha} is set to 1, then instead of a cone, 
the neighbourhood will have the shape of a cylinder.
}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.258.8742&rep=rep1&type=pdf}
}
