% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{get_st_indicators}
\alias{get_st_indicators}
\title{Get spatio-temporal indicators from a data frame
containing spatio-temporal information}
\usage{
get_st_indicators(df, stations_sf, radiuses = c(0.1), stats = c("mean",
  "sd"), alpha = 0.5, neib_type = "cone", time_id = "time",
  site_id = "site_id", vars = c("value"), parallel = FALSE,
  nsplits = 4)
}
\arguments{
\item{df}{A data frame containing spatio-temporal
information}

\item{stations_sf}{An \code{sf} object containing
geographical information on the location of \code{df}}

\item{radiuses}{A vector of values defining the maximum
spatio-temporal distance allowed for an observation to be considered
within a spatio-temporal neighbourhood}

\item{stats}{A vector containing the names of functions
that are to be used to calculate summarizing statistics}

\item{alpha}{a weighting factor for the spatio-temporal distance}

\item{neib_type}{the type of neighborhood to consider.
Can be
\itemize{
\item \code{cone} (default) - a cone with the center of its base at 
the observation (spatial radius growing with time)
\item \code{reversed} - a cone with its peak at the observation
(spatial radius shrinking with time)
}}

\item{time_id}{The name of the column containing 
time-stamps in \code{df}}

\item{site_id}{The name of the column containing 
location IDs in \code{df}}

\item{vars}{The name of the variables to summarize into 
indicators}

\item{parallel}{Boolean indicating whether the code should
run in parallel. Default is FALSE}

\item{nsplits}{Number of subsets of rows to split the data 
frame into so they can be processed in parallel}
}
\value{
A data frame that contains summary statistics
of the values of \code{vars} found within the spatio-temporal 
neighbourhoods of the one or more \code{radiuses} of each pair
 (location ID, time-stamp) in \code{df}
}
\description{
Calculate spatio-temporal indicators of one or more
variables within a spatio-temporal neighborhood of one 
or more maximum radius (in terms of spatio-temporal 
distance).
}
\details{
The spatio-temporal distance is defined as
\deqn{D_{i,j} = d_{i,j} x \alpha + t_{i,j}
x (1-\alpha)}
where d_{i,j} is the spatial distance between
locations, {t_{i,j}} is the temporal distance
between time-stapms and \eqn{\alpha} is a weighting factor.
}
\references{
Ohashi, Orlando, and Luis Torgo. "Wind speed forecasting using spatio-temporal indicators." ECAI. 2012.
}
