% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{get_st_neighbours}
\alias{get_st_neighbours}
\title{Get spatio-temporal neighbourhood}
\usage{
get_st_neighbours(site, time, radius, t_dist_mat, s_dist_mat, alpha,
  time_id = "time", site_id = "site_id")
}
\arguments{
\item{site}{a location ID}

\item{time}{a time-stamp}

\item{radius}{a radius of spatio-temporal distance}

\item{t_dist_mat}{a matrix of normalized temporal distances
between time-stamps (rownames and colnames should be
a concatenation of "TIME_" and the time-stamp)}

\item{s_dist_mat}{a matrix of normalized spatial distances
between locations (rownames and colnames should be
a concatenation of "SITE_" and the location IDs)}

\item{alpha}{a weighting factor for the spatio-temporal distance}

\item{time_id}{the name to give to the column
of time-stamps (Default: time)}

\item{site_id}{the name to give to the column
of location IDs (Default: site_id)}
}
\value{
A data frame where each row describes a neighbour,
with the first two columns containing the location ID and
time-stamp of the central observation, followed by two
columns with the neighbouring location ID and time-stamp,
and a final colunm containing the spatio-temporal distance
between the two.
}
\description{
A function that calculates the observations that
are within a spatio-temporal neighbourhood of a 
certain radius of a time and location.
}
\details{
The spatio-temporal distance is defined as
\deqn{D_{i,j} = d_{i,j} x \alpha + t_{i,j}
x (1-\alpha)}
where d_{i,j} is the spatial distance between
locations, {t_{i,j}} is the temporal distance
between time-stapms and \eqn{\alpha} is a weighting factor.

Note that \code{radius} should always be a number between
zero and \code{min(alpha, alpha-1)}.
Also note that if \code{alpha} is set to 1, then instead of a cone, 
the neighbourhood will have the shape of a cylinder.
}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.258.8742&rep=rep1&type=pdf}
}
