% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_indicators.R
\name{get_time_dist_mat}
\alias{get_time_dist_mat}
\title{Calculate temporal distance matrix}
\usage{
get_time_dist_mat(times, origin = min(times))
}
\arguments{
\item{times}{A vector of time-stamps}

\item{origin}{A date to use as origin for \code{difftime}}
}
\value{
a matrix of distances. Row and column names
are a concatenation of "TIME_" and the time-stamp.
}
\description{
A function that calculates a distance matrix of time-stamps
}
