% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_utils.R
\name{grid_neibs_ord1}
\alias{grid_neibs_ord1}
\title{Get directional neighbours from a regular grid}
\usage{
grid_neibs_ord1(sites, klist)
}
\arguments{
\item{sites}{a matrix or data.frame with site ID equal to row number and 
two columns, corresponding to position in x and y}

\item{klist}{a matrix like the ones returned by consecutive use of 
spdep::dnearneigh and spdep::nblag where a value higher than 0 implies 
that the row and column locations are neighbours}
}
\value{
A matrix with \code{nrow(sites)} rows and four columns containing 
the ID of a location's neighbour in each cardinal direction 
(top, right, bottom and left).
}
\description{
Given a matrix detailing the position of location in a 
regular grid, returns a matrix specifying the IDs of 
immediate neighbours to the right, left, top and bottom 
of each location.
}
