% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_utils.R
\name{lag_multiple_datasets}
\alias{lag_multiple_datasets}
\title{Create a spatio-temporal embed of multiple datasets}
\usage{
lag_multiple_datasets(data_list, LAG_use, SLAGS, min_time = rep(max(LAG_use),
  length(LAG_use)))
}
\arguments{
\item{data_list}{A multi-level list as the ones generated by
\code{generate_multiple_datasets}. The first level corresponds to 
a certain grid size; the second level to a certain time series size
and the third level contains multiple lists containing datasets (that might 
have been generated from different STARMA specifications).}

\item{LAG_use}{A vector of the temporal lag orders to use.}

\item{SLAGS}{A list containing vectors of length of the corresponing 
temporal lag size. Each of the vectors contains values of 0 meaning 
no neighbour columns are used at that temporal lag, and/or values of 1 
meaning that the values of immediate neighbours are included for that
temporal lag.}

\item{min_time}{A vector of the minimum time-stamps to be included in
the datasets. Defaults to \code{rep(max(LAG_use), length(LAG_use)}.}
}
\value{
A list of lists similar to data_list, but with the data sets
having spatio-temporal embeds.
}
\description{
Transform multiple spatio-temporal data sets in a list of lists
into a list of lists containing data frames that have a
spatio-temporal embed. Only data for "interior" points in the grid 
is kept. That is, points that do not have four immediate neighbours 
are filtered out.
}
\seealso{
\code{\link{generate_multiple_datasets}}, 
\code{\link{st_lag_neib_ord1}}
}
