% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nd_kf_xval}
\alias{nd_kf_xval}
\title{Non-dependent cross-validation}
\usage{
nd_kf_xval(data, nfolds, FUN, form, fold.alloc.proc = "Trand_SPrand",
  alloc.pars = NULL, t.buffer = NULL, s.buffer = NULL, s.dists = NULL,
  t.dists = NULL, time = "time", site_id = "site", .keepTrain = TRUE,
  ...)
}
\arguments{
\item{data}{full dataset}

\item{nfolds}{number of folds for the data set to be separated into. \cr
If you would like to set the number of time and space folds separately, 
\code{nfolds} should be set to \code{NULL} and \code{t.nfolds} and
\code{sp.nfolds} should be fed as a list to \code{alloc.pars}
(only available when using \code{fold.alloc.proc} set to 
\code{Tblock_SPchecker}, \code{Tblock_SPcontig} or \code{Tblock_SPrand}).}

\item{FUN}{function with arguments
\itemize{
\item \code{train} training set
\item \code{test} testing set
\item \code{time} column name of time-stamps
\item \code{site_id} column name of location identifiers
\item \code{form} a formula for model learning
\item \code{...} other arguments
}}

\item{form}{a formula for model learning}

\item{fold.alloc.proc}{name of fold allocation function. Should be one of
\itemize{
\item \code{Trand_SPrand} -- each fold contains completely random observations.
The default
\item \code{Tall_SPcontig} - each fold includes all time and a 
contiguous block of space
\item \code{Tall_SPrand} - each fold includes all time and random 
locations in space
\item \code{Tblock_SPrand} -  each fold includes a block of contiguous time
for a randomly assigned part of space
\item \code{Tblock_SPall} - each fold includes a block of contiguous time
for all locations
}}

\item{alloc.pars}{parameters to pass onto \code{fold.alloc.proc}}

\item{t.buffer}{numeric value with the distance of the temporal buffer between
training and test sets. For each instance in the test set, instances that have 
a temporal distance of \code{t.buffer} or less at the same point in space are removed
from the training set.}

\item{s.buffer}{numeric value with the maximum distance of the spatial buffer between
training and test sets. For each instance in the test set, instances that have 
a spatial distance of \code{s.buffer} or less at the same point in time are removed
from the training set.}

\item{s.dists}{a matrix of the distances between the spatial IDs in \code{data}.
The column names and row names should be of type "SITE_<site_id>"}

\item{t.dists}{a matrix of the distances between the time-stamps in \code{data}.
The column names and row names should be of type "TIME_<time>"}

\item{time}{column name of time-stamp in \code{data}. 
Default is "time"}

\item{site_id}{column name of location identifier in \code{data}. 
Default is "site_id"}

\item{.keepTrain}{if TRUE (default), instead of the results of 
\code{FUN} being directly returned, a list is created with both
the results and a \code{data.frame} with the time and site identifiers
of the observations used in the training step.}

\item{...}{other arguments to FUN}
}
\value{
If \code{keepTrain} is \code{TRUE}, a list where each slot
corresponds to one repetition or fold, containing a list with
slots \code{results} containing the results of \code{FUN}, and 
\code{train} containing a data.frame with the \code{time} and 
\code{site_id} identifiers of the observations used in the training
step. Usually, the results of \code{FUN} is a data.frame 
with location identifier \code{site_id}, time-stamp \code{time},
true values \code{trues} and the workflow's predictions
\code{preds}.
}
\description{
Performs a cross-validation experiment where folds can be 
allocated in different ways considering time and/or space
and a certain buffer around the testing set time and/or 
space is removed from the training set.
}
