% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_scale}
\alias{norm_scale}
\title{Feature scaling}
\usage{
norm_scale(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a scaled vector
}
\description{
Normalize values to be within the range between [0,1].
}
