% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{prequential_eval}
\alias{prequential_eval}
\title{Prequential evaluation}
\usage{
prequential_eval(data, nfolds, FUN, form, window = "growing",
  fold.alloc.proc = "Tblock_SPall", alloc.pars = NULL, removeSP = FALSE,
  time = "time", site_id = "site", .keepTrain = TRUE, ...)
}
\arguments{
\item{data}{full dataset}

\item{nfolds}{number of folds for the data set to be separated into. \cr
If you would like to set the number of time and space folds separately, 
\code{nfolds} should be set to \code{NULL} and \code{t.nfolds} and
\code{sp.nfolds} should be fed as a list to \code{alloc.pars}
(only available when using \code{fold.alloc.proc} set to 
\code{Tblock_SPchecker}, \code{Tblock_SPcontig} or \code{Tblock_SPrand}).}

\item{FUN}{function with arguments
\itemize{
\item \code{train} training set
\item \code{test} testing set
\item \code{time} column name of time-stamps
\item \code{site_id} column name of location identifiers
\item \code{form} a formula for model learning
\item \code{...} other arguments
}}

\item{form}{a formula for model learning}

\item{window}{type of blocked-time window ordering considered. 
Should be one of
\itemize{
\item \code{growing} - for each time block being tested, all previous
 time blocks are used for training
\item \code{sliding} - for each time block being tested, the immediately
 previous time blocks are used for training
}}

\item{fold.alloc.proc}{name of fold allocation function. Should be one of
\itemize{
\item \code{Tblock_SPall} - each fold includes a block of contiguous time
for all locations
\item \code{Tblock_SPchecker} - each fold includes a block of contiguous time
for a systematically assigned (checkered) part of space
\item \code{Tblock_SPcontig} - each fold includes a block of contiguous time
for a block of spatially contiguous locations
\item \code{Tblock_SPrand} -  each fold includes a block of contiguous time
for a randomly assigned part of space
}}

\item{alloc.pars}{parameters to pass onto \code{fold.alloc.proc}}

\item{removeSP}{argument that determines whether spatio-temporal blocks
including the space being used for testing should be removed from the training set.
Default is FALSE, meaning the information is not removed}

\item{time}{column name of time-stamp in \code{data}. 
Default is "time"}

\item{site_id}{column name of location identifier in \code{data}. 
Default is "site_id"}

\item{.keepTrain}{if TRUE (default), instead of the results of 
\code{FUN} being directly returned, a list is created with both
the results and a \code{data.frame} with the time and site identifiers
of the observations used in the training step.}

\item{...}{other arguments to FUN}
}
\value{
If \code{keepTrain} is \code{TRUE}, a list where each slot
corresponds to one repetition or fold, containing a list with
slots \code{results} containing the results of \code{FUN}, and 
\code{train} containing a data.frame with the \code{time} and 
\code{site_id} identifiers of the observations used in the training
step. Usually, the results of \code{FUN} is a data.frame 
with location identifier \code{site_id}, time-stamp \code{time},
true values \code{trues} and the workflow's predictions
\code{preds}.
}
\description{
Performs an evaluation procedure where training and test sets can 
be allocated in different ways, while always respecting the ordering 
provided by time (models are trained in the past and tested in the
relative future).
}
