% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics.R
\name{regMetrics}
\alias{regMetrics}
\title{Calculate regression metrics}
\usage{
regMetrics(trues, preds, y_train = NULL, norm = FALSE, aeStatFUN = median,
  seStatFUN = mean, util = FALSE, util.parms = NULL)
}
\arguments{
\item{trues}{a vector of true values}

\item{preds}{a vector of predicted values}

\item{y_train}{a vector of training values}

\item{norm}{a Boolean indicating whether to calculate normalized
regression metrics}

\item{aeStatFUN}{a function to calculate a summary of y_train for 
absolute error normalization. Default is median}

\item{seStatFUN}{a function to calculate a summary of y_train for 
squared error normalization. Default is mean}

\item{util}{a Boolean indicating whether to calculate utility-based
regression metrics}

\item{util.parms}{a named list of parameters to use for calculating utility-based
regression metrics. Should contain slots
\itemize{
\item \code{phi.parms} - the result of function \code{phi.control}
\item \code{phi.control} - if \code{phi.parms} is undefined, then \code{phi.control}
can be provided with a list of named arguments to feed function \code{phi.control} using
\code{y_train}. Default is \code{list(method = "extremes", extr.type="high")}
\item \code{loss.parms} - the results of function \code{loss.control}
\item \code{p} - Default is 0.5
\item \code{thr} - Relevance threshold. Default is 1
\item \code{beta} - Beta for F-measure. Default is 1
}}
}
\value{
a named vector of calculated metrics
}
\description{
Calculate MAE, RMSE and utility-based regression evaluation metrics
}
