% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{responseValues}
\alias{responseValues}
\title{Get response values of a dataset from a formula}
\usage{
responseValues(formula, data, na = NULL)
}
\arguments{
\item{formula}{learning formula}

\item{data}{data set to get the target values from}
}
\value{
A vector of the target values.
}
\description{
Get response values of a dataset from a formula
}
