% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{run_one_experiment}
\alias{run_one_experiment}
\title{Run one in-set/out-set error estimation experiment}
\usage{
run_one_experiment(data, in_set_perc, form, in_estimators, in_est.pars,
  out_estimator = "t_oos", out_est.pars = list(tr.perc = in_set_perc),
  workflow = "simple_workflow", wf.pars = NULL,
  evaluator = "evaluate", eval.pars = NULL, seed = 1234,
  site_id = "site", time = "time")
}
\arguments{
\item{data}{a data frame}

\item{in_set_perc}{a fraction of the data to be used as in-set}

\item{form}{a learning formula}

\item{in_estimators}{a vector of names of estimator functions to use on the in-set data}

\item{in_est.pars}{a named list of the same length as \code{in_estimators} containing lists of
arguments to feed to each estimator apllied to the in-set data}

\item{out_estimator}{the name of the estimator function to use on the out-set data}

\item{out_est.pars}{a list containing arguments to feed to the estimator apllied to the out-set data}

\item{workflow}{the name of the function implementing a workflow}

\item{wf.pars}{a list of arguments to feed to \code{workflow}}

\item{evaluator}{the name of the function to calculate evaluation metrics}

\item{eval.pars}{a list of arguments to feed to \code{evaluator}}

\item{seed}{a seed to set at the start of the experiment}

\item{site_id}{the name of the \code{data} column containing location identifiers}

\item{time}{the name of the \code{data} column containing time-stamps}
}
\value{
A list with two slots: \code{out_estRes} containing the results of the out-set
estimators, and \code{in_estRes} containing a list of results of each estimator used on
in-set data
}
\description{
Run one in-set/out-set error estimation experiment
}
