% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle values/rows}
\usage{
shuffle(x)
}
\arguments{
\item{x}{a vector or data frame}
}
\value{
a vector or data frame
}
\description{
Shuffle the values or rows of a vector or data frame
}
