% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{simple_workflow}
\alias{simple_workflow}
\title{A simple learning and prediction workflow}
\usage{
simple_workflow(train, test, form, model = "lm", handleNAs = NULL,
  min_train = 2, nORp = 0.2, time = "time", site_id = "site", ...)
}
\arguments{
\item{train}{a data frame for training}

\item{test}{a data frame for testing}

\item{form}{a formula describing the model to learn}

\item{model}{the name of the algorithm to use}

\item{handleNAs}{string indicating how to deal with NAs.
If "centralImput", training observations with at least 80\%
of non-NA columns, will have their NAs substituted by the mean
value and testing observatiosn will have their NAs filled in with
mean value regardless.}

\item{min_train}{a minimum number of observations that must be
left to train a model. If there are not enough observations, 
predictions will be \code{NA}. Default is 2.}

\item{nORp}{a maximum number or fraction of columns with missing
values above which a row will be removed from train before 
learning the model. Only works if \code{handleNAs} was
set to centralImputation. Default is 0.2.}

\item{time}{the name of the column in \code{train} and
\code{test} containing time-stamps}

\item{site_id}{the name of the column in \code{train} and
\code{test} containing location IDs}

\item{...}{other parameters to feed to \code{model}}
}
\value{
a data frame containing time-stamps, location IDs,
true values and predicted values
}
\description{
A simple learning and prediction workflow
}
