% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sp_checker}
\alias{sp_checker}
\title{Assign the locations of a regular grid to folds
following a checkered pattern}
\usage{
sp_checker(nfolds, nsites, grid.h = sqrt(nsites), grid.w = sqrt(nsites))
}
\arguments{
\item{nfolds}{number of folds to divide the space into}

\item{nsites}{number of locations in the regular grid}

\item{grid.h}{height of the grid (in number of sites). 
Default is \code{sqrt(nfolds)}}

\item{grid.w}{width of the grid (in number of sites). 
Default is \code{sqrt(nfolds)}}
}
\value{
a vector with the fold assignment of each location
}
\description{
Systematically assigns the locations of a data frame
into folds which are checkered across space for 
cross-validation. Assumes the sites are
sorted (e.g., left to right, bottom to top).
}
