% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sp_contig}
\alias{sp_contig}
\title{Assign the locations of a regular grid to folds
in contiguous square blocks.}
\usage{
sp_contig(nfolds, nsites, grid.h = sqrt(nsites), grid.w = sqrt(nsites))
}
\arguments{
\item{nfolds}{number of folds to divide the space into}

\item{nsites}{number of locations in the regular grid}

\item{grid.h}{height of the grid (in number of sites). 
Default is \code{sqrt(nfolds)}}

\item{grid.w}{width of the grid (in number of sites). 
Default is \code{sqrt(nfolds)}}
}
\value{
a vector with the fold assignment of each location
}
\description{
Assigns the locations of a data frame into contiguous 
blocks folds for cross-validation. Assumes the sites are
 sorted (e.g., left to right, bottom to top).
WARNING: Works well for perfect squares that can be 
divided into \code{nfolds} perfect squares ONLY.
}
