% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_utils.R
\name{st_lag_neib_ord1}
\alias{st_lag_neib_ord1}
\title{Create a spatio-temporal embed of data}
\usage{
st_lag_neib_ord1(data, neibs, p, slags)
}
\arguments{
\item{data}{a matrix of spatio-temporal data where each
row contains the variable information for a time point
and each column corresponds to a certain location}

\item{neibs}{a data frame with information about the
neighbours, like the one returned by \code{grid_neibs_ord1}}

\item{p}{a number for the order of the temporal lag for 
values at a specific location}

\item{slags}{a vector with the length of the order of
temporal lags for values at immediately neighbouring
locations. A value of 0 means no neighbours are used at 
that temporal lag, a value of 1 means that the values
of immediate neighbours are included for that temporal lag.}
}
\value{
a data frame with columns time, site, target, 
as many self_lag columns as \code{p}, and as many
top, bottom, right and left neighbour value lags as the 
sum of \code{slags}.
}
\description{
Given spatio-temporal data, transform it into a data frame
where each row has its own value as target variable and,
as predictors, lagged values recorded in the past at 
its own location and/or in the past at neighbouring 
locations.
}
\seealso{
\code{\link{grid_neibs_ord1}}
}
