% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starma_utils.R
\name{starma_sim}
\alias{starma_sim}
\title{Simulate spatio-temporal data using a STARMA model}
\usage{
starma_sim(model, klist, n, rand.gen = stats::rnorm, innov = NULL,
  seed = NULL, FUN = function(x) {     x }, ...)
}
\arguments{
\item{model}{A list with components \code{ar} and \code{ma}. Each component
contains a matrix of the coefficients where the first index corresponds to the
row and the second index to the column.}

\item{klist}{A list of matrices like the ones returned by consecutive use of 
\code{spdep::dnearneigh} and \code{spdep::nblag} where a value higher than 0
 implies that the row and column locations are neighbours}

\item{n}{The length of the time series to be generated}

\item{rand.gen}{The random generator to be used. Defaults to \code{stats::rnorm}}

\item{innov}{A matrix of initial innovations. Defaults to 
\code{matrix(rand.gen(n*ncol(klist[[1]]), ...), n, ncol(klist[[1]]))}}

\item{seed}{A seed to set before generating the dataset. Defaults to \code{NULL}}

\item{FUN}{A (possibly non-linear) function to apply to the matrix during
STAR data generation.}

\item{...}{Other parameters (?)}
}
\value{
A matrix where each column contains the data for a location,
and each row contains the data for a time-stamp.
}
\description{
Generate a spatio-teporal dataset according to a STARMA model.
}
\references{
See example on page 3 of 
\url{https://cran.r-project.org/web/packages/starma/starma.pdf}
}
