% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starma_utils.R
\name{starma_stat_check}
\alias{starma_stat_check}
\title{Check a STARMA model for stationarity}
\usage{
starma_stat_check(model)
}
\arguments{
\item{model}{A list with components \code{ar} and \code{ma}. Each component
contains a matrix of the coefficients as specified in this functions' details.}
}
\value{
\code{TRUE} if the model is stationary. \code{FALSE}, otherwise.
}
\description{
Checks the coefficients of a STARMA models for stationarity.
}
\details{
The stationarity constraints for a STAR model are the following:
\itemize{
\item STAR(2_11)
  \deqn{- phi_20 + abs(phi_21) <1}
  \deqn{abs(phi_10 + phi_11) < 1 - phi_20 - phi_21}
  \deqn{abs(phi_10 - phi_11) < 1 - phi_20 + phi_21}
\item STAR(2_10), meaning \eqn{phi_21 = 0}
  \deqn{phi_20 > -1}
  \deqn{abs(phi_10) + abs(phi_11) < 1 - phi_20}
\item STAR(2_00), meaning \eqn{phi_21 = 0 & phi_11 = 0}
  \deqn{phi_20 > -1}
  \deqn{abs(phi_10) < 1 - phi_20}
\item STAR(1_1), meaning \eqn{phi_20 = 0 & phi_21 = 0}
  \deqn{abs(phi_10) + abs(phi_11) < 1}
\item STAR(1_0), meaning \eqn{phi=20 = 0 & phi_21 = 0 & phi_11 = 0}
  \deqn{abs(phi_10) < 1}
}
where \code{phi_10 = model$ar[1,1]}, 
\code{phi_11 = model$ar[1,2]},
\code{phi_20 = model$ar[2,1]},
and \code{phi_21 = model$ar[2,2]}

STMA models have the same constraints for theta as the STAR contraints for phi.
}
\references{
\url{http://www.tandfonline.com/doi/abs/10.1080/03610918008812173}
}
