% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{summarize_all_art_exps}
\alias{summarize_all_art_exps}
\title{Summarize the results of all (artificial) in-set/out-set experiments}
\usage{
summarize_all_art_exps(all.res, statFUN, na.rm)
}
\arguments{
\item{all.res}{a multi-level list with where the first level corresponds
to learning model used in the experiment, the second level contains a list
for each grid size of artificial data set, the third level contains a list 
for each time series size. Inside there is a list for each generated set,
the list containing a list of results for each lag embed order}

\item{statFUN}{a function to summarize the evaluation metrics. Default is \code{mean}}

\item{na.rm}{whether to remove NAs in function \code{statFUN}}
}
\value{
A data frame containing columns identifying the learning model,
grid size, time series size, type of STARMA used to generate the data,
order of STARMA used to generate, number of iteration of the generation process
with those settings, lag embed order, gold standard error (that of the out-set),
name of error estimator and estimated error (on the in-set)
}
\description{
Summarize the results of all (artificial) in-set/out-set experiments
}
\seealso{
\code{\link{summarize_multiple_exp}}
}
