% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{sumRes2Tab}
\alias{sumRes2Tab}
\title{Transform a multi-level list of summarized results into a table}
\usage{
sumRes2Tab(sumRes)
}
\arguments{
\item{sumRes}{A multi-level list of summarized results where the first level 
corresponds to learning model used in the experiment, the second level 
contains a list for each grid size of artificial data set, the third level 
contains a list for each time series size, and the next level contains a 
data frame with the results obtained in the out-set (gold-standard or 
"real" error) as well as estimated errors for different estimators
(in wide format)}
}
\value{
A data frame containing columns identifying the learning model,
grid size, time series size, type of STARMA used to generate the data,
order of STARMA used to generate, number of iteration of the generation process
with those settings, lag embed order, gold standard error (that of the out-set),
name of error estimator and estimated error (on the in-set), in long format
}
\description{
Transform a multi-level list of summarized results into a table
}
