% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{t_oos_mc}
\alias{t_oos_mc}
\title{Time-wise Monte Carlo}
\usage{
t_oos_mc(data, tr.perc, ts.perc, nreps, FUN, form, time = "time",
  site_id = "site", .keepTrain = TRUE, ...)
}
\arguments{
\item{data}{full dataset}

\item{tr.perc}{percentage of data used for training.
Remaining will be used for testing}

\item{ts.perc}{percentage of data used for testing}

\item{nreps}{number of repetitions/split-points in experiment}

\item{FUN}{function with arguments
\itemize{
\item \code{train} training set
\item \code{test} testing set
\item \code{time} column name of time-stamps
\item \code{site_id} column name of location identifiers
\item \code{form} a formula for model learning
\item \code{...} other arguments
}}

\item{form}{a formula for model learning}

\item{time}{column name of time-stamp in \code{data}. 
Default is "time"}

\item{site_id}{column name of location identifier in \code{data}. 
Default is "site_id"}

\item{.keepTrain}{if TRUE (default), instead of the results of 
\code{FUN} being directly returned, a list is created with both
the results and a \code{data.frame} with the time and site identifiers
of the observations used in the training step.}

\item{...}{other arguments to FUN}
}
\value{
If \code{keepTrain} is \code{TRUE}, a list where each slot
corresponds to one repetition or fold, containing a list with
slots \code{results} containing the results of \code{FUN}, and 
\code{train} containing a data.frame with the \code{time} and 
\code{site_id} identifiers of the observations used in the training
step. Usually, the results of \code{FUN} is a data.frame 
with location identifier \code{site_id}, time-stamp \code{time},
true values \code{trues} and the workflow's predictions
\code{preds}.
}
\description{
Performs a time-wise Monte Carlo experiment where 
split points are randomly chosen and a window of 
previous observations are used for training, with
a window of following observations used for testing.
}
