% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_alloc.R
\name{Tall_SPcontig}
\alias{Tall_SPcontig}
\title{Spatially blocked CV}
\usage{
Tall_SPcontig(data, nfolds, time = "time", site_id = "site")
}
\arguments{
\item{data}{full dataset}

\item{nfolds}{number of folds}

\item{time}{column name of time-stamp in \code{data}. 
Default is "time"}

\item{site_id}{column name of location identifier in \code{data}. 
Default is "site"}
}
\value{
A list with slots:
\itemize{
\item \code{data}, possibly re-ordered
\item \code{f}, a vector with the fold numbers 
(from \code{1} to \code{nfolds}) of each row in \code{data}
}
}
\description{
Fold allocation of k-fold CV using:
\itemize{
\item all time
\item contiguously blocked locations
}
}
