% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_alloc.R
\name{Tblock_SPrand}
\alias{Tblock_SPrand}
\title{Temporal blocked and randomly assigned spatial CV}
\usage{
Tblock_SPrand(data, nfolds, t.nfolds = round(sqrt(nfolds)),
  sp.nfolds = round(sqrt(nfolds)), time = "time", site_id = "site")
}
\arguments{
\item{data}{full dataset}

\item{nfolds}{number of folds}

\item{t.nfolds}{number of folds across time. Default is sqrt(\code{nfolds})}

\item{sp.nfolds}{number of folds across space. Default is sqrt(\code{nfolds})}

\item{time}{column name of time-stamp in \code{data}. 
Default is "time"}

\item{site_id}{column name of location identifier in \code{data}. 
Default is "site"}
}
\value{
A list with slots:
\itemize{
\item \code{data}, possibly re-ordered
\item \code{f}, a vector with the fold identifiers of each row in \code{data}.
The fold identifier is composed of the concatenation of time-fold number
(from \code{1} to \code{t.nfolds}) and space-fold number
(from \code{1} to \code{sp.nfolds}), separated by "_".
}
}
\description{
Fold allocation of k-fold CV using:
\itemize{
\item blocked time
\item randomly assigned locations
}
}
