% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_alloc.R
\name{Trand_SPrand}
\alias{Trand_SPrand}
\title{Classic k-fold CV}
\usage{
Trand_SPrand(data, nfolds, time = "time", site_id = "site")
}
\arguments{
\item{data}{full dataset}

\item{nfolds}{number of folds}

\item{time}{column name of time-stamp in \code{data}. 
Default is "time"}

\item{site_id}{column name of location identifier in \code{data}. 
Default is "site"}
}
\value{
A list with slots:
\itemize{
\item \code{data}, possibly re-ordered
\item \code{f}, a vector with the fold numbers 
(from \code{1} to \code{nfolds}) of each row in \code{data}
}
}
\description{
Fold allocation of classic k-fold CV:
\itemize{
\item shuffled time
\item shuffled locations
}
}
