% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics.R
\name{mse}
\alias{mse}
\alias{rmse}
\alias{mae}
\alias{nmae}
\alias{nmse}
\alias{nrmse}
\title{Calculate error metrics}
\usage{
mse(y, y_hat, na.rm = TRUE)

rmse(y, y_hat, na.rm = TRUE)

mae(y, y_hat, na.rm = TRUE)

nmae(y, y_hat, y_train = NULL, statFUN = stats::median, na.rm = TRUE)

nmse(y, y_hat, y_train = NULL, statFUN = mean, na.rm = TRUE)

nrmse(y, y_hat, y_train = NULL, statFUN = mean, na.rm = TRUE)
}
\arguments{
\item{y}{a vector of true values}

\item{y_hat}{a vector of predicted values}

\item{na.rm}{boolean indicating whether NAs should be removed. 
Default is TRUE}

\item{y_train}{a vector of training values}

\item{statFUN}{summary statistic to use for normalization. Default is 
median for nmae and mean for nmse.}
}
\value{
one error value
}
\description{
Calculate error metrics
}
\section{Functions}{
\itemize{
\item \code{mse}: mean squared error

\item \code{rmse}: mean squared error

\item \code{mae}: mean absolute error

\item \code{nmae}: normalized mean absolute error

\item \code{nmse}: normalized mean squared error

\item \code{nrmse}: normalized root mean squared error
}}

