% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{realSumRes2Tab}
\alias{realSumRes2Tab}
\title{Transform a multi-level list of summarized results into a table}
\usage{
realSumRes2Tab(sumRes, statFUN = mean, na.rm = FALSE)
}
\arguments{
\item{sumRes}{A multi-level list of summarized results where the first level 
corresponds to learning model used in the experiment, the second level 
contains results for each data set}
}
\value{
A data frame containing columns identifying the learning model,
data set "gold standard"/"real" error/ (that of the out-set), 
name of error estimator and estimated error (on the in-set), in long format
}
\description{
Transform a multi-level list of summarized results into a table
}
